# Clean up & optimize duration calculations