`Process.clock_gettime(Process::CLOCK_MONOTONIC)` is now supported on all platforms and ruby versions we care about. We don’t need a fallback anymore as a result.

### Performance

I've also looked into using `Process.clock_gettime(Process::CLOCK_MONOTONIC, :float_millisecond)` to obtain timestamps in the correct unit directly, so we don't have to do the conversion from seconds ourselves. I've added a benchmark script that gives us several ways to get a timestamp in milliseconds, which is what we need for our duration measurements:

```
Calculating -------------------------------------
Process.clock_gettime in milliseconds (int)
                          4.174M (± 1.8%) i/s -     20.966M in   5.024129s
Process.clock_gettime in milliseconds (float)
                          4.393M (± 2.9%) i/s -     22.114M in   5.038199s
Process.clock_gettime in seconds (float), multiplied by 1000
                          4.526M (± 2.1%) i/s -     22.780M in   5.035625s
Process.clock_gettime in seconds (float), multiplied by 1000.0
                          4.867M (± 2.1%) i/s -     24.461M in   5.028644s
Time.now, multiplied by 1000
                          1.791M (± 1.8%) i/s -      9.045M in   5.051360s

Comparison:
Process.clock_gettime in seconds (float), multiplied by 1000.0:  4866582.5 i/s
Process.clock_gettime in seconds (float), multiplied by 1000:  4525763.2 i/s - 1.08x  slower
Process.clock_gettime in milliseconds (float):  4393428.5 i/s - 1.11x  slower
Process.clock_gettime in milliseconds (int):  4174452.2 i/s - 1.17x  slower
Time.now, multiplied by 1000:  1791272.0 i/s - 2.72x  slower
```

Observations:
- `Time.now` is very slow compared to `clock_gettime`.
- Interestingly, converting ourselves from seconds is slightly faster than asking for the timestamps in milliseconds directly. I was not expecting this
- We can even get a small performance boost to make sure our conversion is Float x Float, rather than Int x Float, by converting our `1000` constant to `1000.0`

I've updated out conversions to use a float constant (`1000.0`).